Migrated from bootstrap to ant-design

+++ src/main/java/tododb/services/TodoDBServiceImpl.java
@@ -3,9 +3,9 @@ package tododb.services;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import tododb.entities.Todo;
+import tododb.entities.TodoStatusEnum;
 import tododb.repositories.TodoRepository;
 
-import java.util.Date;
 import java.util.List;
 
 @Service
@@ -15,12 +15,8 @@ public class TodoDBServiceImpl implements TodoDBService {
     TodoRepository todoRepository;
 
     @Override
-    public List<Todo> findAll(Boolean deleted) {
-        if (deleted) {
-            return todoRepository.findAllDeleted();
-        } else {
-            return todoRepository.findAll();
-        }
+    public List<Todo> findAll(TodoStatusEnum status) {
+        return todoRepository.findAll(status);
     }
 
     @Override
@@ -34,14 +30,18 @@ public class TodoDBServiceImpl implements TodoDBService {
     }
 
     @Override
-    public void deleteTodo(Long todoId) {
-        todoRepository.deleteById(todoId);
+    public void deleteTodo(Long todoId, Boolean softDelete) {
+        if (softDelete) {
+            todoRepository.softDelete(todoId);
+        } else {
+            todoRepository.deleteById(todoId);
+        }
     }
 
     @Override
     public void deleteAll(Boolean softDelete) {
         if (softDelete) {
-            todoRepository.softDeleteAll(new Date());
+            todoRepository.softDeleteAll();
         } else {
             todoRepository.deleteAll();
         }
